Migrate Jira alerter
